TRANSIMS Version 4.0
TransimsNet 4.0.31
A bug was fixed in identifying the area type code for the parking lane-use restrictions found in the link detail file.

LinkSum 4.0.33
The denominator of the relative gap statistics was changed to the shortest path value.

PlanCompare 4.0.47
The denominator of the trip gap statistics was changed to the compare path value.