TRANSIMS Version 5.0
SysLib
The XML control structure was modified to include PARTITIONS=TRUE or PARAMETERS=TRUE in the program section if the program accepts additional command line parameters. 

The control structure associated with partition-ready files now includes .* after the filename ([project_directory].filename.*) in the help/range text of the –C and –X directives.