TRANSIMS Version 4.0
PlanSum 4.0.56
The logic for the output link delay file was modified to copy the travel time from the input link delay file if no volume-delay functions are provided.

PlanCompare 4.0.50
A Trip Gap report was added to the interface and the dominator for the trip gap calculation was changed to the input plan time/cost rather than the compare plan time/cost.