TRANSIMS Version 5.0
SysLib
Trip_Sum_Data class was converted to generate trip length and travel time summary reports for vehicle and passenger trips.

PERIOD_CONTROL_POINT key was moved out of PlanCompare and into Data Service for use by multiple programs.  The control point value was added to the Time_Periods class for global access.

The length units within the Plan data structure were rounded to be compatible with other data structures and reduce the potential for programmer error.

Convert Units method was added to Static Service to enable unit conversion for output reports.

A bug was fixed in the processing method when no subareas are defined and in the look-ahead process of the Simulator object.  Logic was added to exclude traffic control delays from the accumulated stop time used to generate wait time problems. Swapping lane changing behavior was implemented.

Snapshot Output in compressed mode now adds a time step and record number in the index file for the last record in the snapshot.

Filter files were included in the SourceForge project to maintain project subdirectory structures.

PlanSum 5.0.6
The Travel Summary and Passenger Summary reports were implemented using the Trip_Sum_Data class.  The PERIOD_CONTROL_POINT key was added to set the time period for the travel summaries.  The start and end time selection was changed to the plan departure and arrival times.

PlanCompare 5.0.3
The PERIOD_CONTROL_POINT key was moved to Data Service.  The start and end time selection was changed to the plan departure and arrival times.

PlanPrep 5.0.2
The start and end time selection was changed to the plan departure and arrival times.

PlanSelect 5.0.2
The start and end time selection was changed to the plan departure and arrival times.

Microsimulator 5.0.11
A bug was fixed in the processing method when no subareas are defined and in the look-ahead process.  Logic was added to exclude traffic control delays from the accumulated stop time used to generate wait time problems. Swapping lane changing behavior was implemented.

NewFormat 5.0.9
Distance estimates were added to the Plan file conversion.