TRANSIMS Version 5.0 SysLib Data packing methods were added to the trip summary data class to facilitate MPI processing.
The Path Builder plan update option was enhanced to be fully compatible with building the path.  The plan update option now accumulates link flows for updating the link delay file.

PlanSum 5.0.7
MPI processing was added to the trip summary reports.

PathSkim 5.0.3
The distance skims were corrected for the updated path builder.

Router 5.0.7
MPI processing was enabled and the link flows for plan updates was accumulated in order to generate link delay files with total flows and updated travel times.