TRANSIMS Version 5.0 SysLib The Path Builder trace method was corrected for distances between parking lots on the same link.
PathSkim 5.0.5
A bug in the distance skim was corrected for trips that start and end on the same link.  The user selection logic was expanded to enable plan file output from the Origin and Destination Zone Location files.

Router 5.0.8
A bug in the distance skim was corrected for trips that start and end on the same link.