TRANSIMS Version 5.0 SysLib The format of the Plan summary messages written to the print file was modified to simplify GREP functions.
Microsimulator 5.0.15
Logic was added to slow down when approaching the end of a pocket lane or a lane use restrictions.  A lane change attribute was added to the simulation vehicle class to request a lane change in the next time step to avoid conflicts ahead.  The traveler and travel plan data structures were streamlined to improve performance.

LinkSum 5.0.4
The VHT units in the link gap report were corrected.