TRANSIMS Version 4.0
EventSum 4.0.12
Plan updates were limited to completed trips (event files with both start and end time events) and the generalized cost value is updated based on the change in travel time.

Router 4.0.68
INTERPOLATE_LINK_DELAYS key was added to enable interpolate of the link travel time based on the time of day the path enters the link and the mid-points of the two closest time periods in the link delay file.

PlanSum 4.0.57
EXCLUDE_ZERO_VOLUME_LINKS key was added to avoid writing links with zero volume to the output link delay file.

Progression 4.0.7
The logic was changed to avoid updating all signal times when evaluating existing offsets.