TRANSIMS Version 5.0
SysLib
UPDATE_TURN_MOVEMENTS key was added to Flow Time services.  Arrays for storing turning movement data were added when the key is true.

The definition file interface was expanded to include the option for a sort flag in the header records.  Trip and Plan *.def files may now identify the file as traveler or time sorted.

Router 5.0.14
UPDATE_TURNING_MOVEMENTS key was included to accumulate turning movements. Corrections and improvements were made in the way link delay data are isolated in multi-threaded applications.  The sort type of the input trip or plan file is posted on the output plan file.  If the input trip and plan files are not sorted in the same way, an error message is generated.  Applications with both trip and input plan files can now match the records using a time sort.

PlanPrep 5.0.8
The output plan file will be labeled with the known sort information from the input plan file or the value of the PLAN_SORT_TYPE.

TripPrep 5.0.1
The output plan file will be labeled with the known sort information from the input plan file or the value of the PLAN_SORT_TYPE.

PlanSelect 5.0.7
Logic was added to pass the sort type from the input plan file to the output plan file.

PlanCompare 5.0.11
The sort type for the input plan file and the compare file need to be consistent.  The records are matched according to the sort type. Logic was added to pass the sort type from the input plan file to the output plan file.

Microsimulator 5.0.29
The logic for writing the compressed snapshot file was modified to avoid any reference to file positioning.  The compressed record size was changed to 16 bytes and the index record number was changed from signed to unsigned integer to permit larger files (64 gigabytes).