TRANSIMS Version 5.0
Simulator 5.0.30
The Microsimulator was renamed to Simulator to reduce confusion related to simulation methods (macro, meso, and micro).  The macroscopic method was modified to consider passenger car equivalents in the volume and position calculations and reaction time within the queue.  The macroscopic lane assignment was also modified to improve the appearance in the Visualizer.