TRANSIMS Version 5.0
Router 5.0.20
Logic was added to adjust trip schedules and activity durations within tours based the results of previous legs.  Additional logic was included to adjust plan departure times and update vehicle locations when a subset of trips within a household or tour is re-routed.  This is needed to preserve the schedule integrity of the trips and track the location of the vehicle throughout the day.  The memory buffers used to transfer link delay data between MPI machines were modified to avoid transfer buffer limitations for large networks.

PlanSum 5.0.14
The memory buffers used to transfer link delay data between MPI machines were modified to avoid transfer buffer limitations for large networks.