TRANSIMS Version 5.0
SysLib
Minimum and maximum dwell time fields in the vehicle type file were changed for integers to decimal time fields.

Transit run numbers were internally changed from 1 to 0 based values to consistent with other array indices.  

A refinement was made to the data range class to properly round a negative upper range value.  This was needed to avoid problems with the select link option where the BA direction is provided as a negative link number.

The lane use read method was modified to accept 0 as the low end of a vehicle type range.

Router 5.0.22
Changes were made to the lane use restriction logic to properly consider the lane range, record order, and apply, prohibit, and limit control concepts.  The plan processing queue was changed from a queue of individual trips to a queue containing all of the trips made by a household.  The logic to coordinate departures times between trips and tours and adjust activity durations was moved into the Path Builder class.  This change was necessary to avoid vehicle access errors caused by multiple threads simultaneously processing different trips within a household that use the same vehicle.  Logic was also added to reposition the vehicle parking lot for household trips that are not selected for routing.  This permits path building for selected times of day for trips within tours.  The ignore routing problems key was activated to facilitate the processing of trips in a tour after a problem is encountered.

Simulator 5.0.37
Changes were made to the lane use restriction logic to properly consider the record order and the apply, prohibit, and limit control concepts.  The default transit dwell time was changed to the average of the minimum and maximum dwell times coded in the vehicle type file.  Transit run and vehicle numbers were refined.