TRANSIMS Version 5.0
SysLib
Changes were made to the Best Lane Use method within the Path Builder class to return a consistent random delay value for each trip.

The binary write function for 2 and 1 byte integers was enhanced to set the value to the corresponding maximum value when the input integer exceeds the maximum value.

Reset Buffers method was added to the Db_Record class to properly re-initialize the binary record size when the fields are cleared and the new record size is less than the original record size.

Router 5.0.23
Random effects on the lane use delays were made consistent for a given trip.   The lane use “apply” option was modified to ignore pocket lanes.  A bug was fixed in the MPI version when the number of partitions and the number of MPI machines are equal.  A bug was fixed in the binary record length of the new link delay file when the input link delay file includes turning movements, but the output link delay file does not.

Simulator 5.0.38
A sequencing error was fixed for tour simulation.  A bug was fixed in the lane change direction logic was multi-lane use restrictions.  Partition checks and transfers were added to the vehicle loading from parking lots.