TRANSIMS Version 5.0
TransimsNet 5.0.6
Logic was added to include shared lanes from the link detail file in the lane connection procedure.

Simulator 5.0.39
Lane use restrictions with multiple criteria and “required” restriction types were improved.  A unique random number was created for each combination of lane use record, lane number and vehicle to ensure that the random delay is the same each time a given link, lane, traveler, and time of day combination is probed or applied.

Router 5.0.24
The exit check based on the maximum household number was removed when the input trip file is time sorted.