TRANSIMS Version 5.0
TransimsNet 5.0.7
The merge lane logic for Link Detail files was changed to represent the lane entering the link rather than the lane attached to the following link.   If a Link Detail turn or merge lane is specified, the lane length is zero, and a pocket lane warrant is not provided, a default pocket lane length is derived from the link length.  Logic was added to check the combined length of turn and merge pocket lanes on the left and right side of the link to ensure that the lanes do not overlap.

NewFormat 5.0.17
Column numbers for transit schedules were updated to the recent change in run numbers.