TRANSIMS Version 5.0 -- from AECOM branch 1362 September 2011 Release
SysLib
Changes were made to the Best Lane Use method within the Path Builder class to return a consistent random delay value for each trip.

The binary write function for 2 and 1 byte integers was enhanced to set the value to the corresponding maximum value when the input integer exceeds the maximum value.

Reset Buffers method was added to the Db_Record class to properly re-initialize the binary record size when the fields are cleared and the new record size is less than the original record size.

The Db_File logic was modified to detect the Version 4 extension file.t* using the control key file.*

Time-based variable length list control key processing was added.

Router 5.0.27
Random effects on the lane use delays were made consistent for a given trip.   The lane use “apply” option was modified to ignore pocket lanes.  A bug was fixed in the MPI version when the number of partitions and the number of MPI machines are equal.  A bug was fixed in the binary record length of the new link delay file when the input link delay file includes turning movements, but the output link delay file does not.  The exit check based on the maximum household number was removed when the input trip file is time sorted.  A bug was fixed applying incremental travel time updates when an input link delay file is not provided.  Incremental travel time updates within MPI applications were improved. The local access impedance factor model was changed to include a smooth transition and a non-linear impact beyond the local access distance.  WALK_PENALTY_DISTANCES, WALK_PENALTY_FACTOR, BIKE_PENALTY_DISTANCES, BIKE_PENALTY_FACTOR, WAIT_PENALTY_TIMES, and WAIT_PENALTY_FACTOR keys were added to implement the same impedance factoring model to minimize cliff affects.  Initial implementation of transit path building was included.  The code was restructured to improve transit path processing.  The units for the maximum walk and bike distances were corrected. Access to network links for bicycle trips is now assumed to walk legs rather than bike legs.  Access links to nodes and between stops are now included in path building from activity locations and in the middle of paths. A specific random number seed was assigned to each trip.  

Simulator 5.0.40
A sequencing error was fixed for tour simulation.  A bug was fixed in the lane change direction logic for multi-lane use restrictions.  Partition checks and transfers were added to the vehicle loading from parking lots.  Lane use restrictions with multiple criteria and “required” restriction types were improved.  A unique random number was created for each combination of lane use record, lane number and vehicle to ensure that the random delay is the same each time a given link, lane, traveler, and time of day combination is probed or applied.  A bug was fixed in the transit vehicle ID calculation.

TransimsNet 5.0.8
Logic was added to include shared lanes from the link detail file in the lane connection procedure.  The merge lane logic for link detail files was changed to represent the lane entering the link rather than the lane attached to the following link.   If a link detail turn or merge lane is specified, the lane length is zero, and a pocket lane warrant is not provided, a default pocket lane length is derived from the link length.  Logic was added to check the combined length of turn and merge pocket lanes on the left and right side of the link to ensure that the lanes do not overlap.  Logic was added to detect, warn, and correct situations where the angles between two sets of approach and departure links are equal.  The merge lane warrants were flipped so that the from-to facility type concepts are more intuitive.  Output turn penalty file processing was implemented.

TransitNet 5.0.6
Logic was added to generate access links for rail stations.  Column numbers for transit schedules were updated to the recent change in run numbers.  Logic was added to check that the transit time periods match the number of time periods in the Route Nodes file.

NewFormat 5.0.17
Column numbers for transit schedules were updated to the recent change in run numbers.

LocationData 5.0.1
Lookup table indexing was changed from zero based to one based for backward compatibility and more intuitive/simplified scripts.

ArcPlan 5.0.7
Vehicle type and access link file processing were added to support transit route processing and access links in walk and drive trips.  Walk trip path processing was included for all-walk or walk access components of drive and transit trips.  Basic transit access paths were enhanced.