TRANSIMS Version 5.0
SysLib
The Control Service was modified to detect a nested key that starts with a string that equals a previous nested key.

The File Service was modified to print the number of partitions after input files that end with *.

LocationData 5.0.2
A bug was fixed in the subarea zone allocation process.  The new subarea field key now recognizes format attributes.

Router 5.0.28
The walk detail flag was implemented for walk and drive trips.

Simulator 5.0.41
The plan processing was modified to read multiple plan partitions that are independently time sorted.

PathSkim 5.0.12
The skim summary logic was updated to be compatible of the expanded path building procedures.

ExportNet 5.0.0
Initial shell of a new program to export TRANSIMS network data to other file formats.  The first implementation is a UTDF network file for Synchro.