TRANSIMS Version 5.0
SysLib
A bug was fixed in Control Service when writing draw service comment messages to the control file using the –c command line code.

A bug was fixed in Db_File when a non-partitioned file is re-opened as part of a partitioned sequence.

The trip gap reports by time of day were modified to print each time period even if there are no observations.

Logic was added to the Read Link Delay method within Data Service to address the aggregation or disaggregation of time increments that are not even multiples of the input time increment.

NetPrep 5.0.9
The SpdCap lookup table logic was modified to multiply the capacity values by the number of lanes.

Router 5.0.29
Additional logic was added to limit the use of a link to a specific vehicle type range.  A correction was made to the percent problem calculation for MPI implementations.  A bug was fixed when a selection file with only one partition and no input plan file is processed.

ArcNet 5.0.5
The –c option was improved related to draw service comment lines and the default values for time period breaks.

PlanSum 5.0.16
The UPDATE_TURNING_MOVEMENTS key was added to include turning volumes in the output link delay file.