TRANSIMS Version 5.0
ExportNet 5.0.2
NEW_VISSIM_XML_FILE key was added along with signal-related network keys and SUMMARY_TIME_RANGES and SELECT_TIME_OF_DAY to export the TRANSIMS network into the VISSIM XML (ANM) format.

SimSubareas 5.0.3
Logic was added to assign all nodes that are not contained within one of the subarea polygons to a default subarea at the beginning or end of the subarea numbering sequence.

LinkData 5.0.2
The processing logic was expanded to permit multiple volume and/or speed fields to be combined and converted at the same time.  List group options were added to each of the field name fields.