TRANSIMS Version 4.0.8 -- December 2011 Release -- from AECOM Branch 1395
SysLib
A potential divide by zero bug was fixed in the Time Distribution report class.

TransimsNet 4.0.33
A bug was fixed in identifying the area type code for the parking lane-use restrictions found in the link detail file.  A change was made to avoid speed limit and free speed changes to links included in the update list.  A bug was fixed in the right turn pocket lane logic when processing Link Detail files.

LinkSum 4.0.33
The denominator of the relative gap statistics was changed to the shortest path value.

PlanCompare 4.0.50
The denominator of the trip gap statistics was changed to the compare path value. A potential divide by zero bug was fixed in the Time Distribution report class. A divide by zero bug was fixed when the trip duration in the comparison file is zero.  A Trip Gap report was added to the interface and the dominator for the trip gap calculation was changed to the input plan time/cost rather than the compare plan time/cost.

EventSum 4.0.13
The plan update option was corrected to include the leg number in the index.  Plan updates were limited to completed trips (event files with both start and end time events) and the generalized cost value is updated based on the change in travel time.
The New Time Summary File was expanded to include the average absolute difference or error for the trip start time, trip end time, travel time, and schedule variance.  The MINIMUM_SCHEDULE_VARIANCE key was added to define a significant schedule difference.

LinkDelay 4.0.11
PREVIOUS_LINK_FILE key was added to transfer link delay data from a network with different link numbering, but the same node numbers.

PlanSum 4.0.58
The logic for the output link delay file was modified to copy the travel time from the input link delay file if no volume-delay functions are provided.
EXCLUDE_ZERO_VOLUME_LINKS key was added to avoid writing links with zero volume to the output link delay file.  The EXCLUDE_ZERO_VOLUME_LINKS key was replaced by KEEP_INPUT_TRAVEL_TIMES to more clearly define the process of overriding volume-delay equations and keeping the travel times from the input link delay file.

Router 4.0.68
Negative or huge start and arrival times included in trip and activity files are now addressed with warning messages rather than error messages to enable the program to skip the erroneous record and continue processing.  INTERPOLATE_LINK_DELAYS key was added to enable interpolate of the link travel time based on the time of day the path enters the link and the mid-points of the two closest time periods in the link delay file.

Progression 4.0.7
The logic was changed to avoid updating all signal times when evaluating existing offsets.

Validate 4.0.18
Count file processing now averages multiple count records for the same link and time period.

PlanPrep 4.0.16
A bug was fixed using the Delete Household List option.

ConvertTours 4.0.15
Scheduling at-work tours in the same hour as work end time was enhanced by doubly constraining the subtour scheduling to include the return trip start time for the work tour.  Additional scheduling constraints and coordination was added for multiple at-work tours.