TRANSIMS Version 5.0
SysLib
A bug was fixed in the user programming stack report.

The Def_Format method was added as a static member of Db_Header to extract the format type from the def file associated with a filename. 

Db_Matrix class was added for a general matrix DLL interface.  The TPPlus matrix class was reconfigured with Db_Matrix as the base class.  The TransCAD matrix class was added using the Db_Matrix base class.

ExportNet 5.0.7
NEW_ZONE_LOCATION_FILE key was added to write the zone numbers and coordinates with the subarea and at external stations.

ModeChoice 5.0.4
MODE_CONSTANT_FILE was added to input constants using a file in the same format at the calibration output file.  The NEW_CALIBRATION_DATA_FILE was renamed as the NEW_MODE_CONSTANT_FILE for consistency.  These files may or may not include segment records and/or table specific constants.  The mode table field names were changed to the table name included in the input person trip file.  The matrix interface was changed to the generic Db_Matrix class to enable format control using TRIP_FORMAT, NEW_TRIP_FORMAT, and SKIM_FORMAT keys.  The TPPlus and TransCAD matrix formats are the only formats currently implemented.