TRANSIMS Version 4.0
Router 4.0.70
If the mode in an Activity file is carpool/HOV 2, 3 or 4 and the vehicle ID is greater than zero, an HOV vehicle path will now be built for the trip.  Previous logic assumed a carpool mode in the 
Activity file was always an auto passenger trips and converted the mode to magic move.

EventSum 4.0.15
The trip number assigned to the Activity file input was modified to ignore the Activity field value and generate a trip ID in the same way as the Router.

PlanSum 4.0.59
The summary report processing was corrected for plan files that include activity records.