TRANSIMS Version 5.0 SysLib The command line flag –REPORT was added to generate a document file with the program syntax, flag options, control keys, and report labels in a format suitable for quick reference generation.  The flag messages were slightly reordered and _* were replaced with _# for multi-level key help messages.
Logic was added to vehicle ID processing to ensure that the ID value is less than 100.

LineSum 5.0.2
NEW_ACCESS_REPORT_FORMAT and NEW_TOTAL_REPORT_FORMAT keys were added to write output files in user selected format.

Router 5.0.47
A bug was fixed in the multi-threaded processing logic when an input link delay file is not provided.