TRANSIMS Version 5.0 SysLib The vehicle ID validation check was moved into the vehicle and trip file I/O methods.
Router 5.0.48
A bug was fixed in defining the number of file sets when a single partition is processed.

LineSum 5.0.3
The units of measure were added to output definition files and adjusted to accommodate the automatic conversion methods within TRANSIMS.  The units are hard-coded to English at this point in time.  ON_OFF_REPORT_* keys were added to summarize the total, mode, and line boardings and alightings at selected stops.