TRANSIMS Version 5.0 SysLib Get_Bool method was added to the DB Base class.
Logic was added to Execution Service to permit warning messages to be generated when the maximum warning keys are disabled.

Logic was added to the String range class to return low and high values separated by at ‘-‘ in addition to ‘..’.

TransitAccess 5.0.1
New program to generate auto access support links in TPPlus format.

LineSum 5.0.5
Additional logic was added to the Stop and Link reports to expanded functionality.