TRANSIMS Version 5.0
Validate 5.0.9
The period summary flag was not set.

LineSum 5.0.13
SERVICE_FILE, SERVICE_FORMAT, SERVICE_LINE_FIELD, SERVICE_PEAK_FIELD, and SERVICE_OFFPEAK_FIELD keys were added to specify the number of vehicles or train cars servicing the peak and offpeak periods.  Load factors in the AB and BA directions were added to the Link Rider files and the load factor calculation was changed to include the peaking factor.  The LINK_RIDER_PEAK_FACTOR key was changed for hours to a factor (e.g., 1.2) to improve clarity and simplify the service level processing.