TRANSIMS Version 5.0
TripPrep 5.0.7
SYNTHESIZE_VEHICLES key was added to create a new vehicle file from the trip data.

ModeChoice 5.0.11
Protection was added for zones that are not assigned to market segments.

LineSum 5.0.15
A bug was fixed in the Line report when the selected route is the last route in the file.

TransitAccess 5.0.8
ZONE_STATION_MODE_CODES key was added to enable the user to select which station mode codes are given direct walk access links to neighboring zones.

Simulator 5.0.49
If the SIMULATION_END_TIME key is not provided or is set to zero, the default value is defined by the MODEL_END_TIME key rather than 24:00.  Logic was added for pending trips after parking access and vehicle problems.  A bug was fixed for trip ends on boundary links.