TRANSIMS Version 5.0
SysLib
The selection processing logic was simplified to more easily select all or individual persons, tours and trips within a household.

Router 5.0.52
Additional problem messages were included to help identify a more specific reason for a vehicle access error.  A bug was fixed in re-setting the vehicle location when a problem trips within a tour was identified.  The traveler type field was populated in the problem file.

TripPrep 5.0.9
CHECK_ACTIVITY_PATTERNS key was added to perform basic checks of logical sequence of trips times, locations, and vehicles within tours.

ModeChoice 5.0.14
OUTPUT_TRIP_FACTOR key was added to enable the user to multiple the output trip values by a scaling factor.