TRANSIMS Version 5.0 SysLib PAGE_LENGTH key was added to enable the user to override the default page breaks in the middle of a report.
A bug was fixed in the user program while loop and support was added for field numbers in addition to field names.

LineSum 5.0.17
Report processing order is now optionally controlled by the LINESUM_REPORT keys.  The line names in the Stop Report were modified to ignore formatting characters.  The column spacing in the Stop Report was enlarged to minimize number overflow issues.  The transfer row of the Stop Report is now longer printed if transfer modes are not specified.

TripSum 5.0.10
The Check Activity Patterns logic was enhanced to write a problem message in the notes field of the output trip file.  The Synthesize Vehicles logic was enhanced to create a new vehicle for each drive trip whose origin parking location does not match the destination parking location of the previous trip.

FileFormat 5.0.1
COPY_EXISTING_FIELDS, NEW_FILE_HEADER, NEW_DATA_FIELD, and CONVERSION_SCRIPT keys were added to enable field manipulation and calculated fields.