TRANSIMS Version 4.0.9 -- from AECOM branch 1398 December 2012 Release
Router 4.0.70
ADJUST_ACTIVITY_SCHEDULE and IGNORE_ACTIVITY_DURATIONS keys were added to control the start time and duration processing of activity files.  If a trip’s travel time is less than expected, the adjust schedule key will permit the activity to move forward in time, but keep the activity duration the same.  If a trip’s travel time is greater than expected, the ignore durations key will permit the activity duration to be reduced to as little as one minute to enable the next trip to start as close to the original time schedule as possible.  The ignore durations key also drops activity records from the output plan file.  These keys can be applied separately or in combination. If the mode in an Activity file is carpool/HOV 2, 3 or 4 and the vehicle ID is greater than zero, an HOV vehicle path will now be built for the trip.  Previous logic assumed a carpool mode in the Activity file was always an auto passenger trips and converted the mode to magic move.

EventSum 4.0.16
The time summary file processing logic was modified to limit the difference calculations to completed trips and base the calculations on the original trip or activity time schedule rather than the plan file schedule when a trip and/or activity file is provided as input.  The trip number assigned to the Activity file input was modified to ignore the Activity field value and generate a trip ID in the same way as the Router.  Logic was added to correct the trip number generated for Activity files following one person households and when a person starts the day with multiple at-home activities or at some place other than home.

PlanSum 4.0.59
The summary report processing was corrected for plan files that include activity records.

TripPrep 4.0.10
The user program was added to the activity file processing and the logic was modified to apply the user program to both an activity and trip file in the same application.  The random number seed in the user program is now set using the control file.  The Output All and New Household list keys are now active for both trip and activity files in script processing mode.  Selection percentages were added to the activity file processing.  A bug was fixed in the selection logic for activity file time shifts.