TRANSIMS Version 6.0
ArcPlan 6.0.5
Transit path shapes were corrected for paths without walk details and where transfers occur at a common stop.  Additional drawing logic was added to properly display recently added plan details related to access links.  The Num_Legs field was modified to write the leg number rather than the total number of legs.  The Leg_ID field was changed to post the transit route number on transit legs.  All other leg types post the number of links or legs included in the mode.

FileFormat 6.0.2
A bug was fixed in reading the control keys for multiple files.

NewFormat 6.0.7
A bug was fixed in processing a plan file when a trip file is not created.

Router_Service
A bug was fixed in the mode code assigned to a stop that is both the end of one transit leg and the beginning of the next leg for arrival time path building.

Router 6.0.23
Incorporate the Router_Service update.

PathSkim 6.0.6
Incorporate the Router_Service update.