TRANSIMS Version 6.0
LocationData 6.0.2
The PARKING_FILE key was added to avoid warning messages when an Access file is provided.  Logic to link activity locations to transit stops on the same link was added to supplement the connections provided by the Access file for calculating the walk access to transit weight.

ArcPerf 6.0.7
A bug was fixed in setting up the new Arc Performance file.

Simulator_Service
Logic was added to the Macroscopic simulation method to use link capacity as the exit flow rate for the link.  Logic was added to enable some Macroscopic lane changing and to give priority to vehicle loading once the priority loading time parameter is reached.  Transitions from unsimulated to simulated subareas or time periods were improved.

Router 6.0.34
A walk length calculation and the problem counts for file-based output.