TRANSIMS Version 6.0 – June 2014 SysLib A number of improvements were made to the multi-threading classes avoid racing problems and create greater consistency in the method names.  A Compress method was added to the String class to remove all white space from the string.
LocationData 6.0.2
The PARKING_FILE key was added to avoid warning messages when an Access file is provided.  Logic to link activity locations to transit stops on the same link was added to supplement the connections provided by the Access file for calculating the walk access to transit weight.

ArcPerf 6.0.7
A bug was fixed in setting up the new Arc Performance file.
vehicle loading once the priority loading time parameter is reached.

Router 6.0.36
A walk length calculation and the problem counts for file-based output.  Vehicle tracking within a tour was implemented.  Park-n-Ride return trips were forced to return to the same park-n-ride lot.  The PRELOAD_TRANSIT_VEHICLES key was added to enable the program to preload transit vehicles onto each link by time of day based on the current travel times and route schedules.  Preloading is done prior to each feedback iteration.

ArcPlan 6.0.7
A few drawing bugs were fixed for transit plans that do not include walk details.

LineSum 6.0.3
The average trip lengths for aggregate columns were corrected in the line report.

NewLandUse 6.0.4
This is a new program for redistributing population and employment data between zones based on a variety of development density objectives focused primarily on transit-oriented developments and regional activity centers.

TcadImport 6.0.0
This is a new program to read matrix data from a dBase or text file and create the corresponding TransCAD matrix file.

Simulator_Service
Logic was added to the Macroscopic simulation method to use link capacity as the exit flow rate for the link.  Logic was added to enable some Macroscopic lane changing and to give priority to delayed travelers.  Significant changes were made to the multi-threading logic to improve record locking and reduce processing time.  The node-based process approach was replaced