TRANSIMS Version 6.0
TransimsNet 6.0.10
Additional logic was added to identify necessary approach link changes for pocket lanes and lane connections for nodes attached to updated links. 

PlanSum 6.0.8
The performance file output logic was expanded to calculate additional attributes using methods that are more consistent with the Router.

Router 6.0.39
Improvements were made in the DTA and DUE logic to limit path building to trips with significant differences, but also update the travel times for plans that are not re-built.  A bug in the speed calculations for trips that only use a small segment of a link were also corrected.

Simulator 6.0.2
Improvements were made to Macroscopic speed processing and transfers between Unsimulated subareas.  Mesoscopic logic was implemented for lane changing, plan following, random slow downs, traffic controls, and speed calculations.  Output processing and reporting was also improved.